requirements.txt
azure-core==1.26.4
azure-data-tables==12.4.2
aniso8601==9.0.1
asttokens==2.0.8
backcall==0.2.0
bidict==0.22.1
# certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
click==8.1.3
colorama==0.4.5
debugpy==1.6.3
entrypoints==0.4
executing==1.0.0
Faker==18.4.0
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.2
ipykernel==6.15.2
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-core==4.11.1
jupyter_client==7.3.5
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2023.3
# pywin32==304
pyzmq==23.2.1
six==1.16.0
stack-data==0.5.0
tornado==6.2
traitlets==5.3.0
wcwidth==0.2.5
Werkzeug==2.2.3
wincertstore==0.2
